- Refactor the CodeResource to accept only a string to encapsulate the Skema api specifics on the backend - Add comments - Rename event handlers to something better - Move skema proxies into their own package - javadoc comments in server
