+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/CodeRequest.java
@@ -8,14 +8,21 @@ import javax.json.bind.annotation.JsonbProperty;
 import java.io.Serializable;
 import java.util.List;
 
+/**
+ * The CodeRequest instance to send to TA1 for model extraction from text
+ */
 @NoArgsConstructor
 @Data
 @Accessors(chain = true)
 public class CodeRequest implements Serializable {
-	private List<String> files;
+	private List<String> files = List.of("test");
 	private List<String> blobs;
 	@JsonbProperty("system_name")
-	private String systemName;
+	private String systemName = "";
 	@JsonbProperty("root_name")
-	private String rootName;
+	private String rootName = "";
+
+	public CodeRequest(final String code) {
+		blobs = List.of(code);
+	}
 }
