+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/proxies/skema/SkemaProxy.java
@@ -1,7 +1,6 @@
-package software.uncharted.terarium.hmiserver.proxies;
+package software.uncharted.terarium.hmiserver.proxies.skema;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
-import org.keycloak.adapters.spi.HttpFacade;
 import software.uncharted.terarium.hmiserver.models.CodeRequest;
 
 import javax.ws.rs.POST;
@@ -12,6 +11,11 @@ import javax.ws.rs.core.Response;
 
 @RegisterRestClient(configKey = "skema")
 public interface SkemaProxy {
+	/**
+	 * Converts a {@link CodeRequest} to a function network via TA1 Skema
+	 * @param request	the {@link CodeRequest} instance containing the code snippit
+	 * @return	an escaped JSON string of the function network
+	 */
 	@POST
 	@Path("/fn-given-filepaths")
 	@Produces(MediaType.APPLICATION_JSON)
